Move nano::ledger_cache to its own file.
