+++ nano/secure/CMakeLists.txt
@@ -47,6 +47,8 @@ add_library(
   generate_cache_flags.cpp
   ledger.hpp
   ledger.cpp
+  ledger_cache.hpp
+  ledger_cache.cpp
   network_filter.hpp
   network_filter.cpp
   utility.hpp
